+++ src/base.rs
@@ -1,4 +1,5 @@
-use core::{cmp::Ordering, fmt, num::FpCategory};
+use core::cmp::Ordering;
+use core::num::FpCategory;
 
 use hexf::hexf64;
 
@@ -264,109 +265,6 @@ impl TwoFloat {
     }
 }
 
-impl fmt::Display for TwoFloat {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let sign_char = if self.lo().is_sign_positive() {
-            '+'
-        } else {
-            '-'
-        };
-        if f.sign_plus() {
-            match f.precision() {
-                Some(p) => write!(
-                    f,
-                    "{:+.*} {} {:.*}",
-                    p,
-                    self.hi,
-                    sign_char,
-                    p,
-                    self.lo.abs()
-                ),
-                None => write!(f, "{:+} {} {}", self.hi, sign_char, self.lo.abs()),
-            }
-        } else {
-            match f.precision() {
-                Some(p) => write!(f, "{:.*} {} {:.*}", p, self.hi, sign_char, p, self.lo.abs()),
-                None => write!(f, "{} {} {}", self.hi, sign_char, self.lo.abs()),
-            }
-        }
-    }
-}
-
-impl fmt::LowerExp for TwoFloat {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let sign_char = if self.lo().is_sign_positive() {
-            '+'
-        } else {
-            '-'
-        };
-        if f.sign_plus() {
-            match f.precision() {
-                Some(p) => write!(
-                    f,
-                    "{:+.*e} {} {:.*e}",
-                    p,
-                    self.hi,
-                    sign_char,
-                    p,
-                    self.lo.abs()
-                ),
-                None => write!(f, "{:+e} {} {:e}", self.hi, sign_char, self.lo.abs()),
-            }
-        } else {
-            match f.precision() {
-                Some(p) => write!(
-                    f,
-                    "{:.*e} {} {:.*e}",
-                    p,
-                    self.hi,
-                    sign_char,
-                    p,
-                    self.lo.abs()
-                ),
-                None => write!(f, "{:e} {} {:e}", self.hi, sign_char, self.lo.abs()),
-            }
-        }
-    }
-}
-
-impl fmt::UpperExp for TwoFloat {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let sign_char = if self.lo().is_sign_positive() {
-            '+'
-        } else {
-            '-'
-        };
-        if f.sign_plus() {
-            match f.precision() {
-                Some(p) => write!(
-                    f,
-                    "{:+.*E} {} {:.*E}",
-                    p,
-                    self.hi,
-                    sign_char,
-                    p,
-                    self.lo.abs()
-                ),
-                None => write!(f, "{:+E} {} {:E}", self.hi, sign_char, self.lo.abs()),
-            }
-        } else {
-            match f.precision() {
-                Some(p) => write!(
-                    f,
-                    "{:.*E} {} {:.*E}",
-                    p,
-                    self.hi,
-                    sign_char,
-                    p,
-                    self.lo.abs()
-                ),
-                None => write!(f, "{:E} {} {:E}", self.hi, sign_char, self.lo.abs()),
-            }
-        }
-    }
-}
-
 impl PartialEq<f64> for TwoFloat {
     fn eq(&self, other: &f64) -> bool {
         self.hi.eq(other) && self.lo == 0.0
@@ -470,39 +368,6 @@ mod tests {
         assert!(!no_overlap(0.0, 1.0));
         assert!(!no_overlap(0.0, f64::MIN));
         assert!(no_overlap(0.0, 0.0));
-    }
-
-    #[test]
-    fn display_test() {
-        let value = TwoFloat { hi: 1.0, lo: 0.3 };
-        assert_eq!(format!("{}", value), "1 + 0.3");
-        assert_eq!(format!("{}", -value), "-1 - 0.3");
-        assert_eq!(format!("{:+}", value), "+1 + 0.3");
-        assert_eq!(format!("{:.2}", value), "1.00 + 0.30");
-        assert_eq!(format!("{:.2}", -value), "-1.00 - 0.30");
-        assert_eq!(format!("{:+.2}", value), "+1.00 + 0.30");
-    }
-
-    #[test]
-    fn lowerexp_test() {
-        let value = TwoFloat { hi: 1.0, lo: -0.3 };
-        assert_eq!(format!("{:e}", value), "1e0 - 3e-1");
-        assert_eq!(format!("{:e}", -value), "-1e0 + 3e-1");
-        assert_eq!(format!("{:+e}", value), "+1e0 - 3e-1");
-        assert_eq!(format!("{:.2e}", value), "1.00e0 - 3.00e-1");
-        assert_eq!(format!("{:.2e}", -value), "-1.00e0 + 3.00e-1");
-        assert_eq!(format!("{:+.2e}", value), "+1.00e0 - 3.00e-1");
-    }
-
-    #[test]
-    fn upperexp_test() {
-        let value = TwoFloat { hi: 1.0, lo: 0.3 };
-        assert_eq!(format!("{:E}", value), "1E0 + 3E-1");
-        assert_eq!(format!("{:E}", -value), "-1E0 - 3E-1");
-        assert_eq!(format!("{:+E}", value), "+1E0 + 3E-1");
-        assert_eq!(format!("{:.2E}", value), "1.00E0 + 3.00E-1");
-        assert_eq!(format!("{:.2E}", -value), "-1.00E0 - 3.00E-1");
-        assert_eq!(format!("{:+.2E}", value), "+1.00E0 + 3.00E-1");
     }
 
     #[test]