+++ src/functions/fraction.rs
@@ -1,4 +1,5 @@
-use crate::{arithmetic::fast_two_sum, TwoFloat};
+use crate::arithmetic::fast_two_sum;
+use crate::TwoFloat;
 
 impl TwoFloat {
     /// Returns the fractional part of the number.