+++ src/functions/trigonometry.rs
@@ -527,10 +527,9 @@ impl TwoFloat {
 #[cfg(test)]
 mod tests {
     use super::quadrant;
-    use crate::{
-        consts::{FRAC_PI_2, FRAC_PI_4, PI},
-        TwoFloat,
-    };
+
+    use crate::consts::{FRAC_PI_2, FRAC_PI_4, PI};
+    use crate::TwoFloat;
 
     const THRESHOLD: f64 = 1e-10;